More python3 fixes in test harness
Also add variable to Makefile to make it easier to flip back and forth
between python2 and python3

+++ tests/remote-gdb/runtest.py
@@ -278,11 +278,11 @@ def test_read_write_memory(name):
         # Write memory
         for addr, data in tests:
             d.expect('M' + hex(addr)[2:] + ',' +
-                     hex(len(data) / 2)[2:] + ':' + data, 'OK')
+                     hex(int(len(data) / 2))[2:] + ':' + data, 'OK')
 
         # Read and verify
         for addr, data in tests:
-            d.expect('m' + hex(addr)[2:] + ',' + hex(len(data) / 2)[2:], data)
+            d.expect('m' + hex(addr)[2:] + ',' + hex(int(len(data) / 2))[2:], data)
 
         # Try to write a bad address (out of range)
         # Doesn't return an error, test just ensures it
